+++ CHANGELOG.md
@@ -13,11 +13,13 @@ Most recent releases are shown at the top. Each release shows:
 - N/A
 
 ### changed
-- N/A
+- Support `include_reasoning`, `enable_thinking` and `thinking_effort` for
+  Anthropic models on Bedrock
 
 ### fixed:
 - Don't suppy temperature to Anthropic/Bedrock if it is not explicitly supplied. (#255)
 - Fixed bug with GPT-OSS output from AWS GovCloud Bedrock. (#256)
+